requirements.txt
python-telegram-bot[job-queue]==20.6
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
matplotlib==3.8.4
python-dotenv==1.0.1
